ingestion_service/requirements.txt
fastapi
pillow
confluent_kafka
pytesseract
uvicorn